+++ components/StudentStaticSubjectBox.js
@@ -2,6 +2,9 @@ import styles from "../styles/StudentStaticBox.module.css";
 import Image from 'next/image'
 // import subjectIcon from "../images/subject-icon.jpg";
 
+
+//Dynamically get student's tutorials
+
 const StudentStaticSubjectBox = ({ icon, heading, subheading }) => {
   return (
     <div className={styles.staticBox}>
@@ -15,3 +18,17 @@ const StudentStaticSubjectBox = ({ icon, heading, subheading }) => {
 };
 
 export default StudentStaticSubjectBox;
+
+// const StudentStaticSubjectBox = ({ icon, heading, subheading }) => {
+//   return (
+//     <div className={styles.staticBox}>
+//       <Image src={icon} alt="subject-icon" width={360} height={170} />
+//       <div className="d-flex flex-column gap-1">
+//         <b>{heading}</b>
+//         <span>{subheading}</span>
+//       </div>
+//     </div>
+//   );
+// };
+
+// export default StudentStaticSubjectBox;
